+++ src/user-lists/dtos/update-user-list.dto.ts
@@ -1,0 +1,3 @@
+import { CreateUserListDto } from "./create-user-list.dto";
+
+export class UpdateUserListDto extends CreateUserListDto {}