+++ src/workspace/workspace-stats.service.ts
@@ -69,12 +69,14 @@ export class WorkspaceStatsService {
     const queryBuilder = this.baseQueryBuilder();
 
     queryBuilder
+      .withDeleted()
       .leftJoinAndSelect(
         "workspace_repos.repo",
         "workspace_repos_repo",
         "workspace_repos.repo_id = workspace_repos_repo.id"
       )
-      .where("workspace_repos.workspace_id = :id", { id });
+      .where("workspace_repos.deleted_at IS NULL")
+      .andWhere("workspace_repos.workspace_id = :id", { id });
 
     if (options.repos) {
       const sanitizedRepos = sanitizeRepos(options.repos);
@@ -157,12 +159,14 @@ export class WorkspaceStatsService {
     const queryBuilder = this.baseQueryBuilder();
 
     queryBuilder
+      .withDeleted()
       .leftJoinAndSelect(
         "workspace_repos.repo",
         "workspace_repos_repo",
         "workspace_repos.repo_id = workspace_repos_repo.id"
       )
-      .where("workspace_repos.workspace_id = :id", { id });
+      .where("workspace_repos.deleted_at IS NULL")
+      .andWhere("workspace_repos.workspace_id = :id", { id });
 
     const entities = await queryBuilder.getMany();
     const entityRepos = entities.map((entity) => entity.repo.full_name);